drivers: thermal: Notify only thermal zones whose trips are violated
Whenever sensor driver event notification happens, of thermal
framework notifies all thermal zones associated to that sensor for
thermal re-evaluation. In this case, thermal zones whose trips are
not violated are also forced to re-evaluate.

Check if any trip of a thermal zone is violated, then notify that
zone for thermal re-evaluation and ignore other thermal zones
whose trips are not violated.

Change-Id: I3bed9738edc4c1ef4cd3b9748c2b50256941dc7d
Signed-off-by: Manaf Meethalavalappu Pallikunhi <manafm@codeaurora.org>
Signed-off-by: kentalmanis <susukntlmanis88@gmail.com>